requirements.txt
scikit-learn==0.23.1
opencv-python==4.3.0.36
numpy==1.19.0
easydict==1.9
pyyaml==5.3.1